+++ Sources/ExampleCode/ReadmeCode.swift
@@ -128,7 +128,7 @@ public enum ReadmeCode {
     // let response = try await session.respond(to: "What's the weather like in San Francisco?")
     // ...
 
-    for entry in try sessionSchema.decode(session.transcript) {
+    for entry in try sessionSchema.resolve(session.transcript) {
       switch entry {
       case let .toolRun(toolRun):
         switch toolRun {
@@ -163,8 +163,8 @@ public enum ReadmeCode {
 
     print(response.content) // WeatherReport object
 
-    // Access the structured output in the decoded transcript
-    for entry in try sessionSchema.decode(session.transcript) {
+    // Access the structured output in the resolved transcript
+    for entry in try sessionSchema.resolve(session.transcript) {
       switch entry {
       case let .response(response):
         switch response.structuredSegments[0].content {
@@ -211,7 +211,7 @@ public enum ReadmeCode {
     print(response.content)
 
     // Access the input prompt and its groundings separately in the transcript
-    for entry in try sessionSchema.decode(session.transcript) {
+    for entry in try sessionSchema.resolve(session.transcript) {
       switch entry {
       case let .prompt(prompt):
         print(prompt.input) // User input
@@ -274,13 +274,13 @@ public enum ReadmeCode {
 
       // You can also access the generated transcript as it is streamed in
       let transcript = snapshot.transcript
-      let decodedTranscript = try sessionSchema.decode(transcript)
+      let resolvedTranscript = try sessionSchema.resolve(transcript)
 
-      print(transcript, decodedTranscript)
+      print(transcript, resolvedTranscript)
     }
 
     // You can also observe the transcript during streaming
-    for entry in try sessionSchema.decode(session.transcript) {
+    for entry in try sessionSchema.resolve(session.transcript) {
       switch entry {
       case let .response(response):
         switch response.structuredSegments[0].content {
@@ -312,7 +312,7 @@ public enum ReadmeCode {
 
     _ = try session.streamResponse(to: "What's the weather like in San Francisco?")
 
-    for entry in try sessionSchema.decode(session.transcript) {
+    for entry in try sessionSchema.resolve(session.transcript) {
       switch entry {
       case let .toolRun(toolRun):
         switch toolRun {